separate orm to db interface@
