+++ db/db/interface.go
@@ -1,19 +1,19 @@
 package db
 
-import (
-	"gorm.io/driver/mysql"
-	"gorm.io/gorm"
-)
+import "gorm.io/gorm"
 
-type Job struct {
+// Job is, for now, an indeedJob
+type Job indeedJob
+
+type indeedJob struct {
 	gorm.Model
-	Title        string
-	Metadata     string
-	Description  string
-	TimePosted   string
+	Title        string `gorm:"not null"`
+	Metadata     string `gorm:"not null"`
+	Description  string `gorm:"not null"`
+	TimePosted   string `gorm:"not null"`
 	IndeedID     string
 	IndeedURL    string
-	Location     string
+	Location     string `gorm:"not null"`
 	Company      string
 	CompanyImage string
 }
@@ -23,33 +23,6 @@ type JobDatabase interface {
 	AddJob(Job) error
 }
 
-type jobDatabaseImpl struct {
-	*gorm.DB
-}
-
-func (db jobDatabaseImpl) GetJob(id uint) (Job, error) {
-	job := &Job{}
-	result := db.First(job, "id = ?", id)
-	if result.Error != nil {
-		panic(result.Error)
-	}
-
-	return *job, nil
-}
-
-func (db jobDatabaseImpl) AddJob(job Job) error {
-	return nil
-}
-
 func NewJobDatabase() JobDatabase {
-	// refer https://github.com/go-sql-driver/mysql#dsn-data-source-name for details
-	dsn := "root@tcp(127.0.0.1:3306)/prototype?charset=utf8&parseTime=True&loc=Local"
-	db, err := gorm.Open(mysql.Open(dsn), &gorm.Config{})
-	if err != nil {
-		panic(err)
-	}
-
-	db.AutoMigrate(&Job{})
-
-	return jobDatabaseImpl{db}
+	return newJobDatabase()
 }
