+++ db/go.mod
@@ -2,8 +2,6 @@ module db
 
 go 1.17
 
-replace core => ../core
-
 require (
 	github.com/go-sql-driver/mysql v1.6.0 // indirect
 	github.com/jinzhu/inflection v1.0.0 // indirect
